+++ Game/Game/Enemy.cpp
@@ -18,19 +18,19 @@ Enemy::Enemy()
 {
 }
 
-Enemy::Enemy(int x, int y, int width, int height, int size)
+Enemy::Enemy(double x, double y, double width, double height, int size)
 	: Character(x, y, width, height, size)
 {
 
 }
 
-Enemy::Enemy(int x, int y, int size)
+Enemy::Enemy(double x, double y, int size)
 	: Character(x, y, size)
 {
 
 }
 
-Enemy::Enemy(int x, int y)
+Enemy::Enemy(double x, double y)
 	: Character(x, y)
 {
